initialise logger before event manager is used
