+++ src/main/java/uk/gov/ons/census/fwmt/rmadapter/Application.java
@@ -35,24 +35,4 @@ public class Application {
     mapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     return mapper;
   }
-
-  @Value("#{'${logging.profile}' == 'CLOUD'}")
-  private boolean useJsonLogging;
-
-  @PostConstruct
-  public void initJsonLogging() {
-    HashMap<Class<?>, Function<Object, String>> customMappers = new HashMap<>();
-    customMappers.put(LocalTime.class, Object::toString);
-    customMappers.put(LocalDateTime.class, Object::toString);
-
-    LoggingConfigs configs;
-
-    if (useJsonLogging) {
-      configs = LoggingConfigs.builder().customMapper(customMappers).build().useJson();
-    }
-    else {
-      configs = LoggingConfigs.builder().customMapper(customMappers).build();
-    }
-    LoggingConfigs.setCurrent(configs);
-  }
 }
